requirements.txt
arrow==1.2.3
bibtexparser==1.4.0
blessed==1.20.0
chardet==5.1.0
Django==3.2.20
django-crispy-forms==1.14.0
Faker==11.3.0
fontawesome-free==5.15.4
django-environ==0.9.0
django-model-utils==4.3.1
django-picklefield==3.1
django-q==1.3.9
django-settings-export==1.2.1
django-simple-history==3.2.0
django-split-settings==1.2.0
django-sslserver==0.22
django-su==1.0.0
doi2bib==0.4.0
factory-boy==3.2.1
future==0.18.3
humanize==4.6.0
idna==3.4
pyparsing==3.0.9
python-dateutil==2.8.2
python-memcached==1.59
pytz==2022.7.1
pyyaml==6.0.2
redis==3.5.3
requests==2.28.2
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
urllib3==1.26.14
wcwidth==0.2.6
formencode==2.0.1
-r coldfront/plugins/qumulo/requirements.txt